Refactor the way that queries are filtered.

+++ grano/logic/schemata.py
@@ -111,8 +111,12 @@ def save(data, schema=None):
 
 
 def delete(schema):
+    if schema.parent is None:
+        return
     _schema_changed(schema.project.slug, schema.name, 'delete')
-    for attr in schema.attributes:
+    for child in schema.children:
+        child.parent = schema.parent
+    for attr in schema.local_attributes:
         attributes.delete(attr)
     db.session.delete(schema)
